Update
 refactored the coronal, axail and sagittal controllers to inherit from a shared base class that takes a view_type argument.

deleted the duplicate code
